Fix string.prototype methods and add static string methods (#1123)
<!---
Thank you for contributing to Boa! Please fill out the template below, and remove or add any
information as you feel neccesary.
--->

This Pull Request fixes existing string prototype methods in #13 and adds static methods.

It changes the following:

- Fix bugs in existing string prototype methods and improve readability (e.g. rename variables to match the names in spec)
- Add static methods `String.raw`, `String.fromCharCode`, `String.fromCodePoint`
- Fix broken unit tests

+++ boa/src/builtins/regexp/mod.rs
@@ -1760,7 +1760,7 @@ fn advance_string_index(s: &JsString, index: usize, unicode: bool) -> usize {
 
     // 5. Let cp be ! CodePointAt(S, index).
     let (_, offset, _) =
-        crate::builtins::string::code_point_at(s, index as i32).expect("Failed to get code point");
+        crate::builtins::string::code_point_at(s, index as i64).expect("Failed to get code point");
 
     index + offset as usize
 }
